Split features.py into 3 files
